+++ packages/react/src/index.ts
@@ -3,3 +3,4 @@ export * from "./hooks";
 export * from "./components";
 export * from "./engineStore";
 export * from "./dragLayer";
+export * from "./map";